+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+* `dioxus_logger` to `dioxus::logger`
+* update crate: browserinfo(0.1.7)
 
 
 ## [0.1.16] (2026-04-11)